+++ .eslintrc.js
@@ -1,0 +1,4 @@
+module.exports = {
+  extends: ["@olad/eslint-config-typescript"], 
+}
+