+++ packages/eslint-config-basic/test/test.js
@@ -1,0 +1,5 @@
+const a = 123
+
+function add() {
+  return 1 + 2 + 3
+}