+++ packages/eslint-config-typescript/test/index.ts
@@ -1,0 +1,8 @@
+/*
+ * @Date: 2022-04-10 20:29:19
+ * @LastEditors: zhaoxm
+ * @LastEditTime: 2022-04-10 22:23:10
+ * @Description: 
+ */
+
+const a = 123