+++ packages/lint-cli/lib/lint-cli.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = lintCli;
+
+function lintCli() {
+    // TODO
+}