+++ packages/lint-cli/src/types/shared.ts
@@ -1,3 +1,5 @@
+import { PACKAGE_NAMES } from "./../config/consts"
+
 export type PackageJson = {
     name: string,
     version: string,
@@ -5,18 +7,18 @@ export type PackageJson = {
     devDependencies: Recordable
 } & Recordable
 
-export type LintNames = 'eslint' | 'stylelint' | 'commitlint';
+export type lintNameType = keyof typeof PACKAGE_NAMES
 
 export type LintItem = {
-    lintName: LintNames;
-    eslintType?: string;
+    lintName: lintNameType;
+    eslintType?:string
 }
 
 export type Lints = LintItem[];
 
 export type NpmInstallConfig = {
-    packageName: string;
-    templateName: string;
-    targetFileName: string;
-    eslintType: string;
+    packName: string ;
+    // templateName: string;
+    // targetFileName: string;
+    // eslintType: string;
 }