+++ docs/docs/guide/changelog.md
@@ -6,5 +6,5 @@ Feature
 
 - olad-lint : 新增 init 命令，
 - 新增 commitlint-config-olad 依赖包 , 支持单独安转引入
-- 新增 eslint-config-olad 依赖包 ， 支持单独安转引入
+- 新增 @olad/eslint-config 依赖包 ， 支持单独安转引入
 - 新增 stylelint-config-olad 依赖包 ，支持单独安转引入